+++ docs/CHANGELOG.md
@@ -11,8 +11,7 @@
   * Local storage using Go
   * On-device storage now using generic `backends.Backend` api.
   * Improved testing using xla, greatly simplified.
-* Package `graphs`:
-  * Renamed from `graph`: all package names preferring plural names.
+* Package `graph`:
   * Updated `OneHot` to use `Where`.
 * Package `exceptions` removed: using the (same) now external [`github.com/gomlx/exceptions`](github.com/gomlx/exceptions) instead.
 * Package `slices` renamed to `xslices`, not to mix up with the new standard pacakge `slices`.